Change .csv file for action-reward writing
THESE NOTES SHOULD HAVE BEEN INCLUDED ON A PREVIOUS COMMIT!
To add the pose() reward I had to add my own endpoint on the Penguin Pi
itself. Here are the steps for doing this if need be:
1. SSH onto the Penguin Pi, e.g. `ssh pi@192.168.50.1`. The password is
   `PenguinPi`.
2. Open `/usr/bin/penguin-webserver` with a command-line editor like Vim
   or Nano.
3. Include the following function somewhere in the file:
```python
@app.route('/robot/get/pose')
def getencoders():
    global x, y, theta
    return "{},{},{}".format(x, y, theta)
```
Now, hitting a GET request to this endpoint will return the pose.