Add v1 of track edge rewarding
Currently I am applying a really basic heuristic to figure out how well
the Pi is doing of staying on track. It doesn't care about the centre
positioning (to come in future), but it counts the number of gray pixels
in the image as the measure of how well it's doing. This exploits the
simple fact that, in evenly lit conditions, grey pixels have very
similar R, G and B values and none of them are more than about 150 or
so. This is clearly very rough.