Add working live reward plotter
This works by saving a matplotlib figure to file on each time step and
then loading this image as the background of the pygame window. Note
that this is required because in order for key input to work you have to
keep focus on the pygame window and I found that using a simple
plt.show() did not work in this respect.

Also, testing made it very clear that the reward functions are not
properly tuned as I may have liked.